fix: render partial pages for HTMX boost
The Hx-Request header is provided by HTMX to indicate that it is
performing the request as part of a boosted navigation. In this case,
we can send a HTML snippet to be swapped into the HTML body instead of a
full HTML page. HTMX's behavior is the same either way, but this is a
performance optimization to send fewer bytes back to the client when
we're receiving requests from HTMX.